refactor: 모달 구조 분리 및 Sidebar 컴포넌트 정리
